# docs/swagger.yaml
definitions:
  models.Memo:
    properties:
      body:
        type: string
        x-order: "2"
      created_at:
        type: string
        x-order: "3"
      id:
        type: integer
        x-order: "0"
      title:
        type: string
        x-order: "1"
      updated_at:
        type: string
        x-order: "4"
    type: object
info:
  contact: {}
paths:
  /memos:
    get:
      consumes:
      - application/json
      description: Get memos
      operationId: get-memos
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Memo'
            type: array
        "500":
          description: Internal Server Error
      summary: Get memos
    post:
      consumes:
      - application/json
      description: Create a memo
      operationId: post-memo
      parameters:
      - description: Memo
        in: body
        name: memo
        required: true
        schema:
          $ref: '#/definitions/models.Memo'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Memo'
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      summary: Create a memo
  /memos/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a memo
      operationId: delete-memo
      parameters:
      - description: Memo ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Delete a memo
    get:
      consumes:
      - application/json
      description: Get a memo
      operationId: get-memo
      parameters:
      - description: Memo ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Memo'
        "400":
          description: Bad Request
        "404":
          description: Not Found
      summary: Get a memo
    put:
      consumes:
      - application/json
      description: Update a memo
      operationId: put-memo
      parameters:
      - description: Memo ID
        in: path
        name: id
        required: true
        type: integer
      - description: Memo
        in: body
        name: memo
        required: true
        schema:
          $ref: '#/definitions/models.Memo'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Memo'
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Update a memo
swagger: "2.0"